Fix issue with wrong harvest ID being deleted after UNDO
The issue was that when we started a new timer just for toggl, the
harvest ID (other services) were not initialized back to null, while
they should. Those IDs in the workflow configuration file are just
to handle the UNDO functionality and should reflect the status of the
last ran timer through the workflow. If only one service is started,
all the other ones IDs have to be back to null.